+++ sdk-okhttp-client/src/main/java/io/nem/symbol/sdk/infrastructure/okhttp/AccountRepositoryOkHttpImpl.java
@@ -157,9 +157,7 @@ public class AccountRepositoryOkHttpImpl extends AbstractRepositoryOkHttpImpl
                 .map(
                     p ->
                         new AccountLinkVotingKey(
-                            PublicKey.fromHexString(p.getPublicKey()),
-                            (p.getStartEpoch()),
-                            (p.getEndEpoch())))
+                            p.getPublicKey(), (p.getStartEpoch()), (p.getEndEpoch())))
                 .collect(Collectors.toList());
     return new SupplementalAccountKeys(linked, node, vrf, voting);
   }