+++ sdk-vertx-client/src/main/java/io/nem/symbol/sdk/infrastructure/vertx/AccountRepositoryVertxImpl.java
@@ -167,9 +167,7 @@ public class AccountRepositoryVertxImpl extends AbstractRepositoryVertxImpl
                 .map(
                     p ->
                         new AccountLinkVotingKey(
-                            PublicKey.fromHexString(p.getPublicKey()),
-                            (p.getStartEpoch()),
-                            (p.getEndEpoch())))
+                            p.getPublicKey(), (p.getStartEpoch()), (p.getEndEpoch())))
                 .collect(Collectors.toList());
     return new SupplementalAccountKeys(linked, node, vrf, voting);
   }