+++ CHANGELOG.md
@@ -3,7 +3,22 @@ All notable changes to this project will be documented in this file.
 
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [0.22.0] - 3-Dec-2020
+
+## [0.22.1] - 10-Dec-2020
+
+**Milestone**: Catapult-server finality (0.10.0.4)
+ Package  | Version  | Link
+---|---|---
+SDK OkHttp | v0.22.1 | https://repo.maven.apache.org/maven2/io/nem/symbol-sdk-okhttp-client
+SDK Vertx | v0.22.1 | https://repo.maven.apache.org/maven2/io/nem/symbol-sdk-vertx-client
+Catbuffer Library | v0.0.24 | https://repo.maven.apache.org/maven2/io/nem/catbuffer-java
+Client OkHttp | v0.10.5  | https://repo.maven.apache.org/maven2/io/nem/symbol-openapi-okhttp-gson-client
+Client Vertx | v0.10.5  | https://repo.maven.apache.org/maven2/io/nem/symbol-openapi-vertx-client/
+
+- Fixed finalization proof schema version compatibility issue.
+- Account voting key version compatibility issue.
+
+## [0.22.0] - 8-Dec-2020
 
 **Milestone**: Catapult-server finality (0.10.0.4)
  Package  | Version  | Link